하..

+++ src/main/java/com/secondproject/secondproject/service/AssignmentService.java
@@ -137,9 +137,10 @@ public class AssignmentService {
                 .orElseThrow(() -> new EntityNotFoundException("과제 없음"));
         assignment.setTitle(assignSubmitInsertDto.getTitle());
         assignment.setContent(assignSubmitInsertDto.getContent());
+        assignmentRepository.save(assignment);
 
 
-        if (files != null && !files.isEmpty()) {
+        if (files != null && !files.isEmpty() && files.get(0).getSize() > 0) {
             assignmentAttachRepository.deleteByAssignment(assignment);
             for (MultipartFile file : files) {
                 Attachment attachment = attachmentService.save(file, user);
@@ -149,5 +150,21 @@ public class AssignmentService {
                 assignmentAttachRepository.save(assignmentAttach);
             }
         }
+    }
+
+    @Transactional
+    public void deleteAssign(Long assignId) {
+        Assignment assignment = assignmentRepository.findById(assignId)
+                .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 assignment"));
+
+        List<AssignmentAttach> attaches = assignmentAttachRepository.findByAssignment_Id(assignId);
+        for (AssignmentAttach nn : attaches) {
+            assignmentAttachRepository.deleteByAssignment(assignment);
+            attachmentService.deleteById(nn.getAttachment().getId());
+
+        }
+        assignmentRepository.deleteById(assignId);
+
+
     }
 }
